+++ Changelog.md
@@ -1,10 +1,16 @@
 # Changelog
-## 0.0.2
 **If the application crashes copy all files within the update folder into the main folder**
+## 0.0.3
+### Added
+* New Error logger (errors get saved inside the error.txt file)
+### Fixed
+* Fixed a bug, where the Series Updater View was not displayed correctly after an automated series update
+* Fixed a bug, where the Enter Key did not start a series search
+## 0.0.2
 ### Added
 * Show Changelog after Updating
 ### Fixed
-* Fixed a bug where the release date was not displayed correctly
+* Fixed a bug, where the release date was not displayed correctly
 ### Changed
 * Changed the information inside the about tab
 * Moved the dependency locations into the lib folder